Tester om det eksisterer containere, images og volumes for det slettes

+++ Docker/cleandocker.sh
@@ -1,8 +1,16 @@
 #! /bin/bash
 
-docker stop $(docker ps -aq) # Stopp alle kjørende containere
-docker rm $(docker ps -aq) # Slett alle containere
-docker rmi $(docker images -q) # Fjern alle images
-docker volume rm $(docker volume ls -q) # Slett alle volumes
+if [ $(docker ps | wc -l) -gt 1 ]; then
+   docker stop $(docker ps -aq) # Stopp alle kjørende containere
+fi
+if [ $(docker ps -a | wc -l) -gt 1 ]; then
+   docker rm $(docker ps -aq) # Slett alle containere
+fi
+if [ $(docker images | wc -l) -gt 1 ]; then
+   docker rmi $(docker images -q) # Fjern alle images
+fi
+if [ $(docker volume ls | wc -l) -gt 1 ]; then
+   docker volume rm $(docker volume ls -q) # Slett alle volumes
+fi
 
 echo Docker er clean!
